Switched Log In button to input element of type submit
Instead of a button element, the Log In button is now an input element.
The type is submit, so it still submits the form. The log in event is
now attached to the form instead of the button. This allows you to log
in by pressing Enter after typing the password.